Refactored to use whitelist url
